requirements.txt
mysql-connector-python==9.1.0
python-snap7==2.0.2
setuptools==75.1.0
wheel==0.44.0
